changes with classmates
